Simplify backoff for queue websockets client
The websockets client in our queue system implements a manual backoff,
which is no longer necessary with python > 3.6

We remove the test covering the reconnection attempts of the websocket
client used in the job queue. As the code stands now, it is difficult to
test the desired behaviour, cf. the reconnect test in the websockets
library itself. We postpone until we have refactored this part of the
code into an experiment server.