Introduce sleep in job runner
The job runner that runs forward models on compute nodes reports forward
model status as events to the ensemble evaluator.

We are experiencing processing of said events as a bottle-neck. We seek
to reduce the number of events sent by the script, in order to release
some of the stress.

While a forward model is running, the status that is updated is only the
memory usage. It is not necessary to have that detailed stats on memory
usage for longer running processes, so we introduce a sleep time after
an initial phase without sleep.