Let ensemble evaluator monitor reconnect indefinitely
We are experiencing connection trouble between clients and the ensemble
evaluator server. The monitor, used by the GUI, takes down the whole
application when it doesn't manage to reconnect to the server.

We believe in stability of the server - it gets busy, but it will be
able to pick up clients again eventually. So we remove the max number of
reconnects, as done in the job queue.