Disable all websocket ping timeouts
This should help with stability of communication, ironically.
We don't really drop connections, as far as we can tell - rather, the
server gets busy, and cannot respond to pings from clients in a timely
manner. We disable the timeouts, so connections don't get dropped,
hoping that the clients will patiently wait for the server to become
responsive again and receive messages from it.