Rewrite snapshot merging non-recursively
Snapshots contain the state information for realization, steps
and jobs during the course of execution. The existing implememtation
of snapshots was an immutable nested dictionary (pydantic + pyrsistent)
which turned out to be too heavy to handle as number of realizations
and jobs went up, resulting in thread starvation and ultimately hard
crashes.

This rewrite is ad-hoc, only solving what is needed and in an explicit
way. There are no nested dictionaries, instead four top level
dictionaries whose elements are updated using pythons builtin .update().

Co-authored-by: Valentin Krasontovitsch <valentin.krasontovitsch@webstep.no>